add changes to app, searchjobs
