+++ client/src/components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import Auth from '../../utils/auth';
+import Auth from '../utils/auth';
 
 const Header = () => {
   const logout = event => {
