+++ client/src/pages/SearchJobs.js
@@ -1,150 +1,380 @@
-import React, { useState, useEffect } from 'react';
-import { Jumbotron, Container, Col, Form, Button, Card, CardColumns } from 'react-bootstrap';
+import React, { useState } from 'react';
+// import './App.css';
+import $ from 'jquery';
 
-import Auth from '../utils/auth';
-import { searchGoogleJobs } from '../utils/API';
-import { saveJobIds, getSavedJobIds } from '../utils/localStorage';
-//Import useMutation hook from apollo
-import { useMutation } from '@apollo/react-hooks';
-//Import SAVE_BOOK mutation
-import { SAVE_BOOK } from '../utils/mutations';
+ function SearchJobs ()  {
+      const [state, setState]= useState({
+        jobname: '',
+        location: '',
+        country: '',
+        start: 0,
+        end: 25,
+        jobs: []
+   })
+           
+         
+    
 
-const SearchJobs = () => {
-  // create state for holding returned google api data
-  const [searchedJobs, setSearchedJobs] = useState([]);
-  // create state for holding our search field data
-  const [searchInput, setSearchInput] = useState('');
+    function updateState(stateObj){
+        setState({...state ,...stateObj})
+     }
 
-  // create state to hold saved jobId values
-  const [savedJobIds, setSavedJobIds] = useState(getSavedJobIds());
+   function  getJobs(){
+           
+          $.ajax({
+            cache: false,
+            data: $.extend( {publisher: '7778623931867371'
+                    ,v: '2'
+                    ,format: 'json'
+                    ,q: state.jobname
+                    ,co: state.country
+                    ,l: state.location
+                    ,radius: 50
+                    ,limit:25
+                    ,sort: 'date'},
+                    { start: state.start, end: state.nd }
+               ),
+            dataType: 'jsonp',
+            type: 'GET',
+            timeout: 5000,
+            url: 'http://api.indeed.com/ads/apisearch'
+        })
+        .done(function( response ) {
+            updateState({jobs: response.results})
+        })
+     }
 
-  //Imported mutation
-  const [saveJob] = useMutation(SAVE_BOOK);
+    function next(){
+          if (state.end >= 0 && state.start >= 0 && state.jobname && state.country) {
+               setState({...state ,start: state.end})
+               setState({...state ,end: state.end + 25})
+               setTimeout(()=>{
+                    getJobs()
 
-  // set up useEffect hook to save `savedJobIds` list to localStorage on component unmount
-  // learn more here: https://reactjs.org/docs/hooks-effect.html#effects-with-cleanup
-  useEffect(() => {
-    return () => saveJobIds(savedJobIds);
-  });
+               }
+               ,500)
+          }
+          else{
+               setState({...state ,start: 0})
+               setState({...state ,end: 25})
+          }
+     }
 
-  // create method to search for jobs and set state on form submit
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
+  function   prev(){
+          if (state.end >= 0 && state.start >= 25 && state.jobname && state.country) {
+               updateState({...state ,start: state.start - 25})
+               updateState({...state ,end: state.end - 25})
+               setTimeout(()=>{
+                   getJobs()
+               }
+               ,500)
+          }
+          else{
+               setState({...state ,start: 0})
+               setState({...state ,end: 25})
+          }
 
-    if (!searchInput) {
-      return false;
-    }
+     }
+ 
+          return (
+           <div className="App">
+                <div className="search">
+                     <div className="overlay">
+                          <input onChange={ e => updateState({jobname: e.target.value}) } className="jobname" id="jobname" placeholder="jobname" type="text"  name="jobname"/>
+                          <input onChange={ e => updateState({location: e.target.value}) } className="location" id="location" type="text" placeholder="location"  name="location"/>
+                          <select onChange={ e => updateState({country: e.target.value}) } name="country" className="country" id="country">
+                              <option>select country</option>
+                            <option value="AF">Afghanistan</option>
+                            <option value="AX">Åland Islands</option>
+                            <option value="AL">Albania</option>
+                            <option value="DZ">Algeria</option>
+                            <option value="AS">American Samoa</option>
+                            <option value="AD">Andorra</option>
+                            <option value="AO">Angola</option>
+                            <option value="AI">Anguilla</option>
+                            <option value="AQ">Antarctica</option>
+                            <option value="AG">Antigua and Barbuda</option>
+                            <option value="AR">Argentina</option>
+                            <option value="AM">Armenia</option>
+                            <option value="AW">Aruba</option>
+                            <option value="AU">Australia</option>
+                            <option value="AT">Austria</option>
+                            <option value="AZ">Azerbaijan</option>
+                            <option value="BS">Bahamas</option>
+                            <option value="BH">Bahrain</option>
+                            <option value="BD">Bangladesh</option>
+                            <option value="BB">Barbados</option>
+                            <option value="BY">Belarus</option>
+                            <option value="BE">Belgium</option>
+                            <option value="BZ">Belize</option>
+                            <option value="BJ">Benin</option>
+                            <option value="BM">Bermuda</option>
+                            <option value="BT">Bhutan</option>
+                            <option value="BO">Bolivia, Plurinational State of</option>
+                            <option value="BQ">Bonaire, Sint Eustatius and Saba</option>
+                            <option value="BA">Bosnia and Herzegovina</option>
+                            <option value="BW">Botswana</option>
+                            <option value="BV">Bouvet Island</option>
+                            <option value="BR">Brazil</option>
+                            <option value="IO">British Indian Ocean Territory</option>
+                            <option value="BN">Brunei Darussalam</option>
+                            <option value="BG">Bulgaria</option>
+                            <option value="BF">Burkina Faso</option>
+                            <option value="BI">Burundi</option>
+                            <option value="KH">Cambodia</option>
+                            <option value="CM">Cameroon</option>
+                            <option value="CA">Canada</option>
+                            <option value="CV">Cape Verde</option>
+                            <option value="KY">Cayman Islands</option>
+                            <option value="CF">Central African Republic</option>
+                            <option value="TD">Chad</option>
+                            <option value="CL">Chile</option>
+                            <option value="CN">China</option>
+                            <option value="CX">Christmas Island</option>
+                            <option value="CC">Cocos (Keeling) Islands</option>
+                            <option value="CO">Colombia</option>
+                            <option value="KM">Comoros</option>
+                            <option value="CG">Congo</option>
+                            <option value="CD">Congo, the Democratic Republic of the</option>
+                            <option value="CK">Cook Islands</option>
+                            <option value="CR">Costa Rica</option>
+                            <option value="CI">Côte dIvoire</option>
+                            <option value="HR">Croatia</option>
+                            <option value="CU">Cuba</option>
+                            <option value="CW">Curaçao</option>
+                            <option value="CY">Cyprus</option>
+                            <option value="CZ">Czech Republic</option>
+                            <option value="DK">Denmark</option>
+                            <option value="DJ">Djibouti</option>
+                            <option value="DM">Dominica</option>
+                            <option value="DO">Dominican Republic</option>
+                            <option value="EC">Ecuador</option>
+                            <option value="EG">Egypt</option>
+                            <option value="SV">El Salvador</option>
+                            <option value="GQ">Equatorial Guinea</option>
+                            <option value="ER">Eritrea</option>
+                            <option value="EE">Estonia</option>
+                            <option value="ET">Ethiopia</option>
+                            <option value="FK">Falkland Islands (Malvinas)</option>
+                            <option value="FO">Faroe Islands</option>
+                            <option value="FJ">Fiji</option>
+                            <option value="FI">Finland</option>
+                            <option value="FR">France</option>
+                            <option value="GF">French Guiana</option>
+                            <option value="PF">French Polynesia</option>
+                            <option value="TF">French Southern Territories</option>
+                            <option value="GA">Gabon</option>
+                            <option value="GM">Gambia</option>
+                            <option value="GE">Georgia</option>
+                            <option value="DE">Germany</option>
+                            <option value="GH">Ghana</option>
+                            <option value="GI">Gibraltar</option>
+                            <option value="GR">Greece</option>
+                            <option value="GL">Greenland</option>
+                            <option value="GD">Grenada</option>
+                            <option value="GP">Guadeloupe</option>
+                            <option value="GU">Guam</option>
+                            <option value="GT">Guatemala</option>
+                            <option value="GG">Guernsey</option>
+                            <option value="GN">Guinea</option>
+                            <option value="GW">Guinea-Bissau</option>
+                            <option value="GY">Guyana</option>
+                            <option value="HT">Haiti</option>
+                            <option value="HM">Heard Island and McDonald Islands</option>
+                            <option value="VA">Holy See (Vatican City State)</option>
+                            <option value="HN">Honduras</option>
+                            <option value="HK">Hong Kong</option>
+                            <option value="HU">Hungary</option>
+                            <option value="IS">Iceland</option>
+                            <option value="IN">India</option>
+                            <option value="ID">Indonesia</option>
+                            <option value="IR">Iran, Islamic Republic of</option>
+                            <option value="IQ">Iraq</option>
+                            <option value="IE">Ireland</option>
+                            <option value="IM">Isle of Man</option>
+                            <option value="IL">Israel</option>
+                            <option value="IT">Italy</option>
+                            <option value="JM">Jamaica</option>
+                            <option value="JP">Japan</option>
+                            <option value="JE">Jersey</option>
+                            <option value="JO">Jordan</option>
+                            <option value="KZ">Kazakhstan</option>
+                            <option value="KE">Kenya</option>
+                            <option value="KI">Kiribati</option>
+                            <option value="KP">Korea, Democratic Peoples Republic of</option>
+                            <option value="KR">Korea, Republic of</option>
+                            <option value="KW">Kuwait</option>
+                            <option value="KG">Kyrgyzstan</option>
+                            <option value="LA">Lao Peoples Democratic Republic</option>
+                            <option value="LV">Latvia</option>
+                            <option value="LB">Lebanon</option>
+                            <option value="LS">Lesotho</option>
+                            <option value="LR">Liberia</option>
+                            <option value="LY">Libya</option>
+                            <option value="LI">Liechtenstein</option>
+                            <option value="LT">Lithuania</option>
+                            <option value="LU">Luxembourg</option>
+                            <option value="MO">Macao</option>
+                            <option value="MK">Macedonia, the former Yugoslav Republic of</option>
+                            <option value="MG">Madagascar</option>
+                            <option value="MW">Malawi</option>
+                            <option value="MY">Malaysia</option>
+                            <option value="MV">Maldives</option>
+                            <option value="ML">Mali</option>
+                            <option value="MT">Malta</option>
+                            <option value="MH">Marshall Islands</option>
+                            <option value="MQ">Martinique</option>
+                            <option value="MR">Mauritania</option>
+                            <option value="MU">Mauritius</option>
+                            <option value="YT">Mayotte</option>
+                            <option value="MX">Mexico</option>
+                            <option value="FM">Micronesia, Federated States of</option>
+                            <option value="MD">Moldova, Republic of</option>
+                            <option value="MC">Monaco</option>
+                            <option value="MN">Mongolia</option>
+                            <option value="ME">Montenegro</option>
+                            <option value="MS">Montserrat</option>
+                            <option value="MA">Morocco</option>
+                            <option value="MZ">Mozambique</option>
+                            <option value="MM">Myanmar</option>
+                            <option value="NA">Namibia</option>
+                            <option value="NR">Nauru</option>
+                            <option value="NP">Nepal</option>
+                            <option value="NL">Netherlands</option>
+                            <option value="NC">New Caledonia</option>
+                            <option value="NZ">New Zealand</option>
+                            <option value="NI">Nicaragua</option>
+                            <option value="NE">Niger</option>
+                            <option value="NG">Nigeria</option>
+                            <option value="NU">Niue</option>
+                            <option value="NF">Norfolk Island</option>
+                            <option value="MP">Northern Mariana Islands</option>
+                            <option value="NO">Norway</option>
+                            <option value="OM">Oman</option>
+                            <option value="PK">Pakistan</option>
+                            <option value="PW">Palau</option>
+                            <option value="PS">Palestinian Territory, Occupied</option>
+                            <option value="PA">Panama</option>
+                            <option value="PG">Papua New Guinea</option>
+                            <option value="PY">Paraguay</option>
+                            <option value="PE">Peru</option>
+                            <option value="PH">Philippines</option>
+                            <option value="PN">Pitcairn</option>
+                            <option value="PL">Poland</option>
+                            <option value="PT">Portugal</option>
+                            <option value="PR">Puerto Rico</option>
+                            <option value="QA">Qatar</option>
+                            <option value="RE">Réunion</option>
+                            <option value="RO">Romania</option>
+                            <option value="RU">Russian Federation</option>
+                            <option value="RW">Rwanda</option>
+                            <option value="BL">Saint Barthélemy</option>
+                            <option value="SH">Saint Helena, Ascension and Tristan da Cunha</option>
+                            <option value="KN">Saint Kitts and Nevis</option>
+                            <option value="LC">Saint Lucia</option>
+                            <option value="MF">Saint Martin (French part)</option>
+                            <option value="PM">Saint Pierre and Miquelon</option>
+                            <option value="VC">Saint Vincent and the Grenadines</option>
+                            <option value="WS">Samoa</option>
+                            <option value="SM">San Marino</option>
+                            <option value="ST">Sao Tome and Principe</option>
+                            <option value="SA">Saudi Arabia</option>
+                            <option value="SN">Senegal</option>
+                            <option value="RS">Serbia</option>
+                            <option value="SC">Seychelles</option>
+                            <option value="SL">Sierra Leone</option>
+                            <option value="SG">Singapore</option>
+                            <option value="SX">Sint Maarten (Dutch part)</option>
+                            <option value="SK">Slovakia</option>
+                            <option value="SI">Slovenia</option>
+                            <option value="SB">Solomon Islands</option>
+                            <option value="SO">Somalia</option>
+                            <option value="ZA">South Africa</option>
+                            <option value="GS">South Georgia and the South Sandwich Islands</option>
+                            <option value="SS">South Sudan</option>
+                            <option value="ES">Spain</option>
+                            <option value="LK">Sri Lanka</option>
+                            <option value="SD">Sudan</option>
+                            <option value="SR">Suriname</option>
+                            <option value="SJ">Svalbard and Jan Mayen</option>
+                            <option value="SZ">Swaziland</option>
+                            <option value="SE">Sweden</option>
+                            <option value="CH">Switzerland</option>
+                            <option value="SY">Syrian Arab Republic</option>
+                            <option value="TW">Taiwan, Province of China</option>
+                            <option value="TJ">Tajikistan</option>
+                            <option value="TZ">Tanzania, United Republic of</option>
+                            <option value="TH">Thailand</option>
+                            <option value="TL">Timor-Leste</option>
+                            <option value="TG">Togo</option>
+                            <option value="TK">Tokelau</option>
+                            <option value="TO">Tonga</option>
+                            <option value="TT">Trinidad and Tobago</option>
+                            <option value="TN">Tunisia</option>
+                            <option value="TR">Turkey</option>
+                            <option value="TM">Turkmenistan</option>
+                            <option value="TC">Turks and Caicos Islands</option>
+                            <option value="TV">Tuvalu</option>
+                            <option value="UG">Uganda</option>
+                            <option value="UA">Ukraine</option>
+                            <option value="AE">United Arab Emirates</option>
+                            <option value="GB">United Kingdom</option>
+                            <option value="US">United States</option>
+                            <option value="UM">United States Minor Outlying Islands</option>
+                            <option value="UY">Uruguay</option>
+                            <option value="UZ">Uzbekistan</option>
+                            <option value="VU">Vanuatu</option>
+                            <option value="VE">Venezuela, Bolivarian Republic of</option>
+                            <option value="VN">Viet Nam</option>
+                            <option value="VG">Virgin Islands, British</option>
+                            <option value="VI">Virgin Islands, U.S.</option>
+                            <option value="WF">Wallis and Futuna</option>
+                            <option value="EH">Western Sahara</option>
+                            <option value="YE">Yemen</option>
+                            <option value="ZM">Zambia</option>
+                            <option value="ZW">Zimbabwe</option>
+                          </select>
+                          <button  onClick={ () => {
+                                    getJobs()
+                                    updateState({start: 0})
+                                    updateState({end: 25})
+                          }} type="submit" id="searchButton" value="Find Jobs" className="">
+                              <a href="#1"><i className="fa fa-search" aria-hidden="true"></i></a>
+                         </button>
+                     </div>
 
-    try {
-      const response = await searchGoogleJobs(searchInput);
+                     </div>
 
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
 
-      const { items } = await response.json();
 
-      const jobData = items.map((job) => ({
-        jobId: job.id,
-        authors: job.volumeInfo.authors || ['No author'],
-        title: job.volumeInfo.title,
-        description: job.volumeInfo.description,
-        image: job.volumeInfo.imageLinks?.thumbnail || '',
-      }));
+                <div className="results">
 
-      setSearchedJobs(jobData);
-      setSearchInput('');
-    } catch (err) {
-      console.error(err);
-    }
-  };
+                     <div>
+                          <button className={ state.start ? "" : "hidden" } onClick={ () => prev()}>prev</button>
+                          <button className={ state.jobs.length > 0 ? "" : "hidden" } onClick={ () => next()}>next</button>
+                     </div>
 
-  // create function to handle saving a job to our database
-  const handleSaveJob = async (jobId) => {
-    // find the job in `searchedJobs` state by the matching id
-    const jobSaving = searchedJobs.find((job) => job.jobId === jobId);
+                     {
+                          state.jobs.map( (value,index)=>{
+                               return(
+                                    <div key={index} className="job">
+                                        <h4  > <a href={value.url} target="_blank"> TITLE : {value.jobtitle} </a> </h4>
+                                        <h4  > Location : {value.city}, {value.country} </h4>
+                                        <h4  > DATE : {value.date} </h4>
+                                        <h4  > POSTED : {value.formattedRelativeTime} </h4>
+                                        <h4  > DESCRIPTION : {value.snippet} </h4>
+                                    </div>
+                               )
+                          })
+                     }
 
-    // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
 
-    if (!token) {
-      return false;
-    }
-
-    try {
-      // imported SAVE_BOOK mutation
-      await saveJob({ variables: { job: jobSaving } });
-
-      // if job successfully saves to user's account, save job id to state
-      setSavedJobIds([...savedJobIds, jobSaving.jobId]);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  return (
-    <>
-      <Jumbotron fluid className='text-light bg-dark'>
-        <Container>
-          <h1>Search for Jobs!</h1>
-          <Form onSubmit={handleFormSubmit}>
-            <Form.Row>
-              <Col xs={12} md={8}>
-                <Form.Control
-                  name='searchInput'
-                  value={searchInput}
-                  onChange={(e) => setSearchInput(e.target.value)}
-                  type='text'
-                  size='lg'
-                  placeholder='Search for a job'
-                />
-              </Col>
-              <Col xs={12} md={4}>
-                <Button type='submit' variant='success' size='lg'>
-                  Submit Search
-                </Button>
-              </Col>
-            </Form.Row>
-          </Form>
-        </Container>
-      </Jumbotron>
-
-      <Container>
-        <h2>
-          {searchedJobs.length
-            ? `Viewing ${searchedJobs.length} results:`
-            : 'Search for a job to begin'}
-        </h2>
-        <CardColumns>
-          {searchedJobs.map((job) => {
-            return (
-              <Card key={job.jobId} border='dark'>
-                {job.image ? (
-                  <Card.Img src={job.image} alt={`The cover for ${job.title}`} variant='top' />
-                ) : null}
-                <Card.Body>
-                  <Card.Title>{job.title}</Card.Title>
-                  <p className='small'>Authors: {job.authors}</p>
-                  <Card.Text>{job.description}</Card.Text>
-                  {Auth.loggedIn() && (
-                    <Button
-                      disabled={savedJobIds?.some((savedJobId) => savedJobId === job.jobId)}
-                      className='btn-block btn-info'
-                      onClick={() => handleSaveJob(job.jobId)}>
-                      {savedJobIds?.some((savedJobId) => savedJobId === job.jobId)
-                        ? 'This job has already been saved!'
-                        : 'Save this Job!'}
-                    </Button>
-                  )}
-                </Card.Body>
-              </Card>
-            );
-          })}
-        </CardColumns>
-      </Container>
-    </>
-  );
-};
+                </div>
+           </div>
+          );
+   
+}
 
 export default SearchJobs;
+
