+++ server/config/connection.js
@@ -2,9 +2,8 @@ const mongoose = require('mongoose');
 
 mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost/IndeedAPI', {
   useNewUrlParser: true,
-  useUnifiedTopology: true,
-  useCreateIndex: true,
-  useFindAndModify: false,
+  useUnifiedTopology: true
+ 
 });
 
 module.exports = mongoose.connection;
